reformat all types to match
