+++ examples/transactions_example.rs
@@ -141,7 +141,7 @@ async fn main() -> Result<(), Box<dyn Error>> {
 
     let payment_response = match client.send_payment(payment_payload, private_key).await {
         Ok(payment_response) => {
-            println!("{}", payment_response);
+            println!("{:?}", payment_response);
             println!("Payment transaction sent successfully");
             payment_response
         }
@@ -195,10 +195,13 @@ async fn main() -> Result<(), Box<dyn Error>> {
     for attempt in 1..=5 {
         println!("   Attempt {}/5: Checking transaction status...", attempt);
 
-        match client.get_transaction_by_hash(tx_hash).await {
+        match client
+            .get_transaction_by_hash(&format!("{:?}", tx_hash))
+            .await
+        {
             Ok(tx) => {
                 println!("Transaction confirmed on chain:");
-                println!("{}", tx);
+                println!("{:?}", tx);
                 confirmed = true;
                 break;
             }
@@ -220,7 +223,10 @@ async fn main() -> Result<(), Box<dyn Error>> {
 
     // Try to get transaction receipt for additional details
     println!("\nFetching transaction receipt...");
-    match client.get_transaction_receipt_by_hash(tx_hash).await {
+    match client
+        .get_transaction_receipt_by_hash(&format!("{:?}", tx_hash))
+        .await
+    {
         Ok(receipt) => {
             println!("Transaction receipt:");
             println!("{}", receipt);
