+++ src/api/mod.rs
@@ -10,12 +10,8 @@ pub mod transactions;
 // Re-export client types from the new client module
 pub use crate::client::{Client, ClientBuilder, Network};
 
-// Re-export commonly used API types
-pub use accounts::{AccountQuery, TokenAccountQuery};
-pub use tokens::{
-    BlacklistAction, BlacklistTokenRequest, BurnTokenRequest, MintTokenRequest, PauseAction,
-    PauseTokenRequest, TokenAuthorityPayload, TokenAuthorityRequest, TokenBlacklistPayload,
-    TokenBurnPayload, TokenMetadataUpdatePayload, TokenMintPayload, TokenPausePayload,
-    TokenWhitelistPayload, UpdateMetadataRequest, WhitelistAction, WhitelistTokenRequest,
+// Re-export commonly used API types now from types module
+pub use crate::requests::{
+    PaymentPayload, TokenAuthorityPayload, TokenBlacklistPayload, TokenBurnPayload,
+    TokenMetadataUpdatePayload, TokenMintPayload, TokenPausePayload, TokenWhitelistPayload,
 };
-pub use transactions::{FeeEstimateRequest, PaymentPayload, PaymentRequest, TransactionReceipt};
