+++ src/api/states.rs
@@ -3,7 +3,7 @@
 use crate::client::Client;
 use crate::client::config::api_path;
 use crate::client::config::endpoints::states::LATEST_EPOCH_CHECKPOINT;
-use crate::{LatestStateResponse, Result};
+use crate::{Result, responses::LatestStateResponse};
 
 impl Client {
     /// Get the latest epoch and checkpoint information.
@@ -48,8 +48,6 @@ mod tests {
         let state = LatestStateResponse {
             epoch: 123,
             checkpoint: 456,
-            checkpoint_hash: "0xabcdef1234567890".to_string(),
-            checkpoint_parent_hash: "0x1234567890abcdef".to_string(),
         };
 
         let json = serde_json::to_string(&state).unwrap();
@@ -57,10 +55,5 @@ mod tests {
 
         assert_eq!(state.epoch, deserialized.epoch);
         assert_eq!(state.checkpoint, deserialized.checkpoint);
-        assert_eq!(state.checkpoint_hash, deserialized.checkpoint_hash);
-        assert_eq!(
-            state.checkpoint_parent_hash,
-            deserialized.checkpoint_parent_hash
-        );
     }
 }
