+++ src/api/transactions.rs
@@ -5,169 +5,15 @@ use crate::client::config::api_path;
 use crate::client::config::endpoints::transactions::{
     BY_HASH, ESTIMATE_FEE, PAYMENT, RECEIPT_BY_HASH,
 };
-use crate::crypto::Signable;
 use crate::crypto::sign_transaction_payload;
-use crate::{Hash, Result, Signature, Transaction};
-use alloy_primitives::{Address, U256};
-use alloy_primitives::{B256, keccak256};
+use crate::requests::{FeeEstimateRequest, PaymentPayload, PaymentRequest};
+use crate::responses::FeeEstimate;
+use crate::responses::TransactionReceipt;
+use crate::{Hash, Result, Transaction};
+use alloy_primitives::U256;
 #[cfg(test)]
 use rlp::encode as rlp_encode;
-use rlp::{Encodable, RlpStream};
-use serde::{Deserialize, Serialize};
-use std::fmt::{Display, Formatter, Result as FmtResult};
-
-/// Payment transaction payload.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct PaymentPayload {
-    /// Recent epoch number.
-    pub recent_epoch: u64,
-    /// Recent checkpoint number.
-    pub recent_checkpoint: u64,
-    /// Chain ID.
-    pub chain_id: u64,
-    /// Account nonce.
-    pub nonce: u64,
-    /// Recipient address.
-    pub recipient: Address,
-    /// Amount to transfer.
-    #[serde(serialize_with = "serialize_token_amount_decimal")]
-    pub value: U256,
-    /// Token address (use native token address for native transfers).
-    pub token: Address,
-}
-
-/// Serialize U256 as decimal string instead of hex (L1 compatibility).
-fn serialize_token_amount_decimal<S>(
-    value: &U256,
-    serializer: S,
-) -> std::result::Result<S::Ok, S::Error>
-where
-    S: serde::Serializer,
-{
-    serializer.serialize_str(&value.to_string())
-}
-
-impl Display for PaymentPayload {
-    fn fmt(&self, f: &mut Formatter<'_>) -> FmtResult {
-        write!(
-            f,
-            "Payment to {}: value {}, token {}, nonce {}, epoch {}, checkpoint {}, chain {}",
-            self.recipient,
-            self.value,
-            self.token,
-            self.nonce,
-            self.recent_epoch,
-            self.recent_checkpoint,
-            self.chain_id
-        )
-    }
-}
-
-impl Encodable for PaymentPayload {
-    fn rlp_append(&self, s: &mut RlpStream) {
-        s.begin_list(7);
-        s.append(&self.recent_epoch);
-        s.append(&self.recent_checkpoint);
-        s.append(&self.chain_id);
-        s.append(&self.nonce);
-        s.append(&self.recipient.as_slice());
-        // Encode U256 as compact bytes (no leading zeros) to match L1
-        let value_bytes = self.value.to_be_bytes_vec();
-        let mut compact_bytes = value_bytes;
-        while !compact_bytes.is_empty() && compact_bytes[0] == 0 {
-            compact_bytes.remove(0);
-        }
-        if compact_bytes.is_empty() {
-            compact_bytes = vec![0];
-        }
-        s.append(&compact_bytes);
-        s.append(&self.token.as_slice());
-    }
-}
-
-impl PaymentPayload {
-    /// Calculate the signature hash for this payload.
-    /// This matches the L1 implementation's signature_hash method.
-    pub fn signature_hash(&self) -> B256 {
-        let encoded = rlp::encode(self);
-        keccak256(&encoded)
-    }
-}
-
-impl Signable for PaymentPayload {
-    fn signature_hash(&self) -> B256 {
-        PaymentPayload::signature_hash(self)
-    }
-}
-
-/// Payment transaction request.
-#[derive(Debug, Clone, Serialize)]
-pub struct PaymentRequest {
-    #[serde(flatten)]
-    pub payload: PaymentPayload,
-    /// Signature for the payload.
-    pub signature: Signature,
-}
-
-/// Fee estimation request.
-#[derive(Debug, Clone, Serialize)]
-pub struct FeeEstimateRequest {
-    /// From address.
-    pub from: Address,
-    /// Value to transfer.
-    pub value: Option<U256>,
-    /// Token address (optional).
-    pub token: Option<Address>,
-}
-
-/// Transaction receipt response.
-#[derive(Debug, Clone, Deserialize)]
-pub struct TransactionReceipt {
-    /// If transaction is executed successfully.
-    pub success: bool,
-    /// Transaction Hash.
-    pub transaction_hash: String,
-    /// Index within the block.
-    pub transaction_index: Option<u64>,
-    /// Hash of the checkpoint this transaction was included within.
-    pub checkpoint_hash: Option<String>,
-    /// Number of the checkpoint this transaction was included within.
-    pub checkpoint_number: Option<u64>,
-    /// Fee used.
-    pub fee_used: u128,
-    /// Address of the sender.
-    pub from: String,
-    /// Address of the receiver. None when its a contract creation transaction.
-    pub to: Option<String>,
-    /// Token address created, or None if not a deployment.
-    pub token_address: Option<String>,
-}
-
-impl Display for TransactionReceipt {
-    fn fmt(&self, f: &mut Formatter<'_>) -> FmtResult {
-        writeln!(f, "Transaction Receipt:")?;
-        writeln!(f, "  Success: {}", self.success)?;
-        writeln!(f, "  Transaction Hash: {}", self.transaction_hash)?;
-        writeln!(f, "  Fee Used: {}", self.fee_used)?;
-        if let Some(idx) = self.transaction_index {
-            writeln!(f, "  Transaction Index: {}", idx)?;
-        }
-        if let Some(hash) = &self.checkpoint_hash {
-            writeln!(f, "  Checkpoint Hash: {}", hash)?;
-        }
-        if let Some(num) = self.checkpoint_number {
-            writeln!(f, "  Checkpoint Number: {}", num)?;
-        }
-        writeln!(f, "  From: {}", self.from)?;
-        if let Some(to) = &self.to {
-            writeln!(f, "  To: {}", to)?;
-        }
-        if let Some(token) = &self.token_address {
-            write!(f, "  Token Address: {}", token)?;
-        }
-        Ok(())
-    }
-}
+use serde::Deserialize;
 
 impl Client {
     /// Send a payment transaction.
@@ -282,7 +128,7 @@ impl Client {
     ///     Ok(())
     /// }
     /// ```
-    pub async fn estimate_fee(&self, request: FeeEstimateRequest) -> Result<crate::FeeEstimate> {
+    pub async fn estimate_fee(&self, request: FeeEstimateRequest) -> Result<FeeEstimate> {
         let mut path = ESTIMATE_FEE.to_string();
         let mut query_params = Vec::new();
 
@@ -311,7 +157,7 @@ impl Client {
             .parse::<u128>()
             .map_err(|_| crate::Error::custom("Invalid fee format from API".to_string()))?;
 
-        Ok(crate::FeeEstimate {
+        Ok(FeeEstimate {
             gas_limit: 21000, // Default gas limit for simple transactions
             gas_price: U256::from(fee_amount / 21000), // Calculated gas price
             total_fee: U256::from(fee_amount),
@@ -355,6 +201,7 @@ impl Client {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use alloy_primitives::Address;
     use std::str::FromStr;
 
     #[test]
