+++ src/types/mod.rs
@@ -1,17 +1,37 @@
 //! Type definitions for the OneMoney SDK.
 
-pub mod accounts;
-pub mod checkpoints;
+// Original SDK types
 pub mod common;
-pub mod states;
 pub mod tokens;
-pub mod transactions;
 
-// Re-export commonly used types
-pub use accounts::*;
-// Note: chains module is currently empty
-pub use checkpoints::*;
+// New organized API types
+pub mod requests;
+pub mod responses;
+
+// L1-compatible API signature types
+pub mod api_signature;
+
+// Re-export commonly used types from original SDK
 pub use common::*;
-pub use states::*;
-pub use tokens::*;
-pub use transactions::*;
+// Note: accounts, checkpoints, states, transactions types are now in responses/
+
+// Re-export original token types (avoid conflicts with API types)
+pub use tokens::{Authority, AuthorityAction};
+
+// Re-export action types from requests module
+pub use requests::{BlacklistAction, PauseAction, WhitelistAction};
+
+// Re-export L1-compatible API types (these take precedence)
+pub use api_signature::*;
+
+// Re-export request types
+pub use requests::tokens::*;
+pub use requests::transactions::*;
+
+// Re-export response types
+pub use responses::accounts::*;
+pub use responses::chains::*;
+pub use responses::checkpoints::*;
+pub use responses::states::*;
+pub use responses::tokens::*;
+pub use responses::transactions::*;
